requirements.txt
flask
flask-socketio
pymongo
flask-bcrypt
python-dotenv
flask-session
gunicorn
